requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
audioop-lts==0.2.2
ccxt==4.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
discord==2.3.2
discord.py==2.6.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pycares==4.10.0
pycparser==2.22
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
yarl==1.20.1
